app/requirements.txt
# --- PDF stack harmonization ---
pdfplumber==0.11.7
pdfminer.six==20250506
fastapi
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.1
minio==7.2.8
qdrant-client==1.9.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
tenacity==8.5.0
python-multipart
pymupdf==1.24.9
mammoth==1.6.0
beautifulsoup4
lxml==5.3.0
Pillow
openai>=1.30.0
backoff>=2.2.1
python-dateutil==2.9.0.post0
llama-index-core>=0.11.7,<0.12.0
llama-index-llms-openai==0.2.3
jsonschema==4.23.0
boto3==1.34.162
paddleocr>=2.7.0
# CPU-only PaddlePaddle (explicit pin; paddleocr may install if missing)
paddlepaddle>=2.6.0
mineru[core]>=2.5.4
docling>=0.15.0
markdown
circuitbreaker
